Many changes/refactorings too long to describe, The focus was on developing the stepped version with steps on removal and modification of free space too.

+++ GlyphAtlasSrc/Glyph/Glyph.h
@@ -6,7 +6,7 @@
 
 struct Glyph
 {
-    uint16 textureId{};
+    uint8 textureId{};
     Rect rect;
 
     void Update(const std::string& serialized)
